Cleanup and finish first implementation
