+++ apps/hops-fileshare/main/fileshare/infrastructure/ApplicationServices.kt
@@ -2,26 +2,23 @@ package fileshare.infrastructure
 
 import fileshare.domain.FileSharingService
 import io.ktor.client.HttpClient
-import io.ktor.config.ApplicationConfig
-import java.net.URL
+import io.ktor.client.features.json.JsonFeature
+import io.ktor.client.features.json.serializer.KotlinxSerializer
+import kotlinx.serialization.json.Json
 
 class ApplicationServices(
-    applicationConfig: ApplicationConfig,
+    applicationConfig: Config,
 ) {
     val fileSharingService: FileSharingService
 
-    private val baseHttpClient = HttpClient()
+    private val baseHttpClient = HttpClient() {
+        install(JsonFeature) {
+            serializer = KotlinxSerializer(Json { ignoreUnknownKeys = true })
+        }
+    }
 
     init {
-        val config = Config.FileStore(
-            URL(applicationConfig.property("fileStore.baseUrl").getString()),
-            applicationConfig.property("fileStore.bucketName").getString(),
-            applicationConfig.property("fileStore.requiresAuth").getString().toBooleanStrict(),
-            URL(applicationConfig.property("fileStore.tokenFetchUrl").getString()),
-            applicationConfig.property("fileStore.virusScanningEnabled").getString().toBooleanStrict(),
-            URL(applicationConfig.property("fileStore.virusScannerUrl").getString()),
-            applicationConfig.property("fileStore.unScannedBucketName").getString(),
-        )
+        val config = applicationConfig.fileStore
         val gcpHttpTransport = GCPHttpTransport(baseHttpClient, config)
         val virusScanner = HttpVirusScanner(
             gcpHttpTransport,
