+++ apps/hops-fileshare/main/fileshare/infrastructure/GCPHttpTransport.kt
@@ -7,8 +7,6 @@ import io.ktor.client.features.ClientRequestException
 import io.ktor.client.features.auth.Auth
 import io.ktor.client.features.auth.providers.BearerTokens
 import io.ktor.client.features.auth.providers.bearer
-import io.ktor.client.features.json.JsonFeature
-import io.ktor.client.features.json.serializer.KotlinxSerializer
 import io.ktor.client.request.get
 import io.ktor.client.request.headers
 import io.ktor.client.request.post
@@ -23,7 +21,6 @@ import io.ktor.utils.io.ByteReadChannel
 import java.util.Base64
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.Json
 
 class GCPHttpTransport(private val baseHttpClient: HttpClient, private val config: Config.FileStore) {
     private val httpClient: HttpClient
@@ -31,21 +28,12 @@ class GCPHttpTransport(private val baseHttpClient: HttpClient, private val confi
     init {
         httpClient = baseHttpClient.config {
             if (config.requiresAuth) {
-                val tokenClient = baseHttpClient.config {
-                    install(JsonFeature) {
-                        serializer = KotlinxSerializer(Json { ignoreUnknownKeys = true })
-                    }
-                }
                 install(Auth) {
                     bearer {
-                        loadTokens { fetchToken(tokenClient) }
-                        refreshTokens { fetchToken(tokenClient) }
+                        loadTokens { fetchToken(baseHttpClient) }
+                        refreshTokens { fetchToken(baseHttpClient) }
                     }
                 }
-            }
-
-            install(JsonFeature) {
-                serializer = KotlinxSerializer(Json { ignoreUnknownKeys = true })
             }
         }
     }
